requirements.txt
# Runtime Modules
fire==0.5.0
# Dev & Testing Modules
black==22.12.0
flake8==4.0.1
flake8-black==0.3.5
isort==5.11.0
pillow==9.4.0
image==1.5.33